Add logo, move file
